@@ -20,7 +20,7 @@
         }
     },
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "ext-json": "*",
         "ext-zlib": "*",
         "aws/aws-sdk-php": "^3.209",
