@@ -7,8 +7,11 @@ namespace Keboola\ObjectEncryptor\Tests;
 use Keboola\ObjectEncryptor\EncryptorOptions;
 use Keboola\ObjectEncryptor\Exception\ApplicationException;
 use Keboola\ObjectEncryptor\Exception\UserException;
+use Keboola\ObjectEncryptor\Wrapper\BranchTypeProjectGKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentAKVWrapper;
+use Keboola\ObjectEncryptor\Wrapper\ComponentGKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ComponentKMSWrapper;
+use Keboola\ObjectEncryptor\Wrapper\GkmsClientFactory;
 use Keboola\ObjectEncryptor\Wrapper\KmsClientFactory;
 
 class ComponentWrapperTest extends AbstractTestCase
@@ -24,19 +27,22 @@ class ComponentWrapperTest extends AbstractTestCase
     }
 
     /**
-     * @return ComponentKMSWrapper[][]|ComponentAKVWrapper[][]
+     * @return ComponentAKVWrapper[][]|ComponentGKMSWrapper[][]|ComponentKMSWrapper[][]
      */
     public function wrapperProvider(): array
     {
+        putenv('GOOGLE_APPLICATION_CREDENTIALS=' . getenv('TEST_GOOGLE_APPLICATION_CREDENTIALS'));
+
         $kmsOptions = new EncryptorOptions(
             stackId: 'some-stack',
             kmsKeyId: self::getkmsKeyId(),
             kmsRegion: self::getkmsRegion(),
             backoffMaxTries: 1,
         );
-        $componentWrapperKMS = new ComponentKMSWrapper(
-            (new KmsClientFactory())->createClient($kmsOptions),
-            $kmsOptions,
+        $gkmsOptions = new EncryptorOptions(
+            stackId: 'some-stack',
+            gkmsKeyId: self::getGkmsKeyId(),
+            backoffMaxTries: 1,
         );
 
         $componentWrapperAKV = new ComponentAKVWrapper(new EncryptorOptions(
@@ -44,19 +50,31 @@ class ComponentWrapperTest extends AbstractTestCase
             akvUrl: self::getAkvUrl(),
             backoffMaxTries: 1,
         ));
+        $componentWrapperGKMS = new ComponentGKMSWrapper(
+            (new GkmsClientFactory())->createClient($kmsOptions),
+            $gkmsOptions,
+        );
+
+        $componentWrapperKMS = new ComponentKMSWrapper(
+            (new KmsClientFactory())->createClient($kmsOptions),
+            $kmsOptions,
+        );
 
         return [
-            'KMS' => [
-                $componentWrapperKMS,
-            ],
             'AKV' => [
                 $componentWrapperAKV,
+            ],
+            'GKMS' => [
+                $componentWrapperGKMS,
+            ],
+            'KMS' => [
+                $componentWrapperKMS,
             ],
         ];
     }
 
     /**
-     * @param ComponentKMSWrapper|ComponentAKVWrapper $wrapper
+     * @param ComponentAKVWrapper|ComponentGKMSWrapper|ComponentKMSWrapper $wrapper
      * @dataProvider wrapperProvider
      */
     public function testEncrypt($wrapper): void
@@ -66,6 +84,65 @@ class ComponentWrapperTest extends AbstractTestCase
         $encrypted = $wrapper->encrypt($secret);
         self::assertNotEquals($secret, $encrypted);
         self::assertEquals($secret, $wrapper->decrypt($encrypted));
+    }
+
+    public function testEncryptDifferentStackAKV(): void
+    {
+        $wrapper = new ComponentAKVWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            akvUrl: self::getAkvUrl(),
+            backoffMaxTries: 1,
+        ));
+        $wrapper->setComponentId('dummy-component');
+        $secret = 'mySecretValue';
+
+        $encrypted = $wrapper->encrypt($secret);
+        self::assertNotEquals($secret, $encrypted);
+        self::assertEquals($secret, $wrapper->decrypt($encrypted));
+
+        $wrapper = new ComponentAKVWrapper(new EncryptorOptions(
+            stackId: 'some-other-stack',
+            akvUrl: self::getAkvUrl(),
+            backoffMaxTries: 1,
+        ));
+        $wrapper->setComponentId('dummy-component');
+        self::expectException(UserException::class);
+        self::expectExceptionMessage('Deciphering failed.');
+        $wrapper->decrypt($encrypted);
+    }
+
+    public function testEncryptDifferentStackGKMS(): void
+    {
+        $options = new EncryptorOptions(
+            stackId: 'some-stack',
+            gkmsKeyId: self::getGkmsKeyId(),
+            backoffMaxTries: 1,
+        );
+
+        $wrapper = new ComponentGKMSWrapper(
+            (new GkmsClientFactory())->createClient($options),
+            $options,
+        );
+        $wrapper->setComponentId('dummy-component');
+        $secret = 'mySecretValue';
+
+        $encrypted = $wrapper->encrypt($secret);
+        self::assertNotEquals($secret, $encrypted);
+        self::assertEquals($secret, $wrapper->decrypt($encrypted));
+
+        $options = new EncryptorOptions(
+            stackId: 'some-other-stack',
+            gkmsKeyId: self::getGkmsKeyId(),
+            backoffMaxTries: 1,
+        );
+        $wrapper = new ComponentGKMSWrapper(
+            (new GkmsClientFactory())->createClient($options),
+            $options,
+        );
+        $wrapper->setComponentId('dummy-component');
+        self::expectException(UserException::class);
+        self::expectExceptionMessage('Deciphering failed.');
+        $wrapper->decrypt($encrypted);
     }
 
     public function testEncryptDifferentStackKMS(): void
@@ -104,33 +181,8 @@ class ComponentWrapperTest extends AbstractTestCase
         $wrapper->decrypt($encrypted);
     }
 
-    public function testEncryptDifferentStackAKV(): void
-    {
-        $wrapper = new ComponentAKVWrapper(new EncryptorOptions(
-            stackId: 'some-stack',
-            akvUrl: self::getAkvUrl(),
-            backoffMaxTries: 1,
-        ));
-        $wrapper->setComponentId('dummy-component');
-        $secret = 'mySecretValue';
-
-        $encrypted = $wrapper->encrypt($secret);
-        self::assertNotEquals($secret, $encrypted);
-        self::assertEquals($secret, $wrapper->decrypt($encrypted));
-
-        $wrapper = new ComponentAKVWrapper(new EncryptorOptions(
-            stackId: 'some-other-stack',
-            akvUrl: self::getAkvUrl(),
-            backoffMaxTries: 1,
-        ));
-        $wrapper->setComponentId('dummy-component');
-        self::expectException(UserException::class);
-        self::expectExceptionMessage('Deciphering failed.');
-        $wrapper->decrypt($encrypted);
-    }
-
     /**
-     * @param ComponentKMSWrapper|ComponentAKVWrapper $wrapper
+     * @param ComponentAKVWrapper|ComponentGKMSWrapper|ComponentKMSWrapper $wrapper
      * @dataProvider wrapperProvider
      */
     public function testEncryptDifferentComponent($wrapper): void
@@ -145,6 +197,33 @@ class ComponentWrapperTest extends AbstractTestCase
         self::expectException(UserException::class);
         self::expectExceptionMessage('Deciphering failed.');
         $wrapper->decrypt($encrypted);
+    }
+
+    public function testInvalidSetupAKV(): void
+    {
+        self::expectException(ApplicationException::class);
+        self::expectExceptionMessage('Cipher key settings are invalid.');
+        new ComponentAKVWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            kmsKeyId: 'some-key',
+            kmsRegion: 'some-region',
+        ));
+    }
+
+    public function testInvalidSetupGKMS(): void
+    {
+        $options = new EncryptorOptions(
+            stackId: 'some-stack',
+            kmsKeyId: 'some-key',
+            kmsRegion: null,
+        );
+
+        self::expectException(ApplicationException::class);
+        self::expectExceptionMessage('Cipher key settings are invalid.');
+        new ComponentGKMSWrapper(
+            (new GkmsClientFactory())->createClient($options),
+            $options,
+        );
     }
 
     public function testInvalidSetupKMS(): void
@@ -163,19 +242,8 @@ class ComponentWrapperTest extends AbstractTestCase
         );
     }
 
-    public function testInvalidSetupAKV(): void
-    {
-        self::expectException(ApplicationException::class);
-        self::expectExceptionMessage('Cipher key settings are invalid.');
-        new ComponentAKVWrapper(new EncryptorOptions(
-            stackId: 'some-stack',
-            kmsKeyId: 'some-key',
-            kmsRegion: 'some-region',
-        ));
-    }
-
     /**
-     * @param ComponentKMSWrapper|ComponentAKVWrapper $wrapper
+     * @param ComponentAKVWrapper|ComponentGKMSWrapper|ComponentKMSWrapper $wrapper
      * @dataProvider wrapperProvider
      */
     public function testInvalidSetupEncryptComponent($wrapper): void
@@ -186,7 +254,7 @@ class ComponentWrapperTest extends AbstractTestCase
     }
 
     /**
-     * @param ComponentKMSWrapper|ComponentAKVWrapper $wrapper
+     * @param ComponentAKVWrapper|ComponentGKMSWrapper|ComponentKMSWrapper $wrapper
      * @dataProvider wrapperProvider
      */
     public function testInvalidSetupDecryptComponent($wrapper): void
